Api controllers parameters validation implemented.

+++ src/LMS/Areas/Api/Controllers/AreasController.cs
@@ -30,18 +30,34 @@ namespace LMS.Areas.Api.Controllers
         [HttpGet("{id}")]
         public UserAreaVM Get(string id, bool includeGoals = true, bool onlyLastGoals = true)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Area id is not set");
+            }
             return _userAreaService.Get(id, new AreaListOptions { IncludeGoals = includeGoals, OnlyLastGoals = onlyLastGoals });
         }
 
         [HttpPost]
         public UserAreaVM Post([FromBody]UserAreaVM userArea)
         {
+            if (!ModelState.IsValid)
+            {
+                throw new ArgumentException();
+            }
             return _userAreaService.Add(userArea);
         }
 
         [HttpPut("{id}")]
         public UserAreaVM Put(string id, [FromBody]UserAreaVM userArea)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Area id is not set");
+            }
+            if (!ModelState.IsValid)
+            {
+                throw new ArgumentException();
+            }
             userArea.Id = id;
             return _userAreaService.Update(userArea);
         }
@@ -49,6 +65,10 @@ namespace LMS.Areas.Api.Controllers
         [HttpDelete("{id}")]
         public string Delete(string id)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Area id is not set");
+            }
             _userAreaService.Delete(id);
             return id;
         }
