+++ src/LMS/Areas/Api/Controllers/CalendarTasksController.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using LMS.Areas.Api.ViewModels;
 using LMS.Services;
 using Microsoft.AspNet.Mvc;
@@ -21,6 +22,14 @@ namespace LMS.Areas.Api.Controllers
         [HttpPost("goals/{goalid}/calendartasks")]
         public CalendarTaskVM Post([FromBody]CalendarTaskVM task, string goalid)
         {
+            if (string.IsNullOrEmpty(goalid))
+            {
+                throw new ArgumentException("Goalid is not set");
+            }
+            if (!ModelState.IsValid)
+            {
+                throw new ArgumentException();
+            }
             return _calendarTasksService.Add(task, goalid);
         }
 
@@ -28,6 +37,10 @@ namespace LMS.Areas.Api.Controllers
         [HttpDelete("calendartasks/{id}")]
         public string Delete(string id)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Calendar task id is not set");
+            }
             _calendarTasksService.Delete(id);
             return id;
         }
