+++ src/LMS/Areas/Api/Controllers/GoalsController.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using LMS.Areas.Api.ViewModels;
 using LMS.Services;
 using Microsoft.AspNet.Mvc;
@@ -19,6 +20,10 @@ namespace LMS.Areas.Api.Controllers
         [HttpGet("areas/{areaid}/goals")]
         public IEnumerable<GoalVM> GetList(string areaid, bool includeCompletedGoals = true, bool onlyLastGoals = true)
         {
+            if (string.IsNullOrEmpty(areaid))
+            {
+                throw new ArgumentException("Areaid is not set");
+            }
             var options = new GoalListOptions
             {
                 AreaId = areaid,
@@ -32,12 +37,24 @@ namespace LMS.Areas.Api.Controllers
         [HttpGet("goals/{id}")]
         public GoalVM Get(string id, bool includeGoals = true, bool onlyLastGoals = true)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Goal id is not set");
+            }
             return _goalService.Get(id);
         }
 
         [HttpPost("areas/{areaid}/goals")]
         public GoalVM Post([FromBody]GoalVM goal, string areaid)
         {
+            if (string.IsNullOrEmpty(areaid))
+            {
+                throw new ArgumentException("Areaid id is not set");
+            }
+            if (!ModelState.IsValid)
+            {
+                throw new ArgumentException();
+            }
             return _goalService.Add(goal, areaid);
         }
 
@@ -45,6 +62,14 @@ namespace LMS.Areas.Api.Controllers
         [HttpPut("goals/{id}")]
         public GoalVM Put(string id, [FromBody]GoalVM userArea)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Goal id is not set");
+            }
+            if (!ModelState.IsValid)
+            {
+                throw new ArgumentException();
+            }
             userArea.Id = id;
             return _goalService.Update(userArea);
         }
@@ -53,8 +78,13 @@ namespace LMS.Areas.Api.Controllers
         [HttpDelete("goals/{id}")]
         public string Delete(string id)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                throw new ArgumentException("Goal id is not set");
+            }
             _goalService.Delete(id);
             return id;
         }
     }
 }
+
